Synchronize dev with the latest main branch while preserving dev-only changes.

+++ Modules/Cleaning/app/Models/CleaningBookingWorkerAssignment.php
@@ -68,10 +68,13 @@ final class CleaningBookingWorkerAssignment extends Model
             'worker_finished_property_rooms' => 'array',
             'room_count' => 'integer',
             'rooms_weight' => 'decimal:2',
-            'service_share_amount' => 'integer',
-            'travel_fee' => 'integer',
-            'admin_margin_amount' => 'integer',
-            'worker_amount' => 'integer',
+            // These columns are DECIMAL(12,2). Casting them to integer silently
+            // truncated room-weight shares such as 166.67 to 166 in API/domain
+            // calculations. Preserve the stored precision instead.
+            'service_share_amount' => 'float',
+            'travel_fee' => 'float',
+            'admin_margin_amount' => 'float',
+            'worker_amount' => 'float',
         ];
     }
